% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobitGwas.R
\name{extract_genotype_columns}
\alias{extract_genotype_columns}
\title{Extracts columns from raw genotype files that we actually need for regression}
\usage{
extract_genotype_columns(genotype_table)
}
\arguments{
\item{genotype_table}{The genotype file parsed as a \link[data.table:data.table]{data.table::data.table}}
}
\value{
A \link[data.table:data.table]{data.table::data.table} with only IDs and dosages remaining
}
\description{
Extracts columns from raw genotype files that we actually need for regression
}
