% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobitGwas.R
\name{outer_join}
\alias{outer_join}
\title{Implements an outer join using \link[base:merge]{base::merge}.}
\usage{
outer_join(x, y)
}
\arguments{
\item{x, y}{The data frames to join.}
}
\value{
An outer join of a and b
}
\description{
We need this because \link[base:funprog]{base::Reduce} can't pass additional arguments to its function.
}
