% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobitGwas.R
\name{rank_transform}
\alias{rank_transform}
\title{Rank Transformation for Tobit Regression}
\usage{
rank_transform(
  values,
  detection_limit = min(values, na.rm = TRUE),
  na_is_below_dt = FALSE
)
}
\arguments{
\item{values}{The raw values to transform}

\item{detection_limit}{The detection limit of the method used to measure the values}

\item{na_is_below_dt}{Indicates whether all NAs measurements below the detection limit.}
}
\value{
A named list with two elements.
\code{values} is a vector of transformed measurements.
\code{detection_limit} is the transformed detection limit.
}
\description{
Rank transforms values that do not fit a normal distribution to fit a
normal distribution and calculates the detection limit of the transformed data.
}
\details{
If \code{detection_limit} is not given, the minimum observed value is assumed
to be the detection limit

If \code{na_is_below_dt} is TRUE, all NAs are assumed to be measurements below
detection and will be set to the detection limit before transformation.
If FALSE, NAs are not modified and it is assumed that all measurements
below detection limit have been set to a value less than or equal to the
detection limit.
}
