% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobitGwas.R
\name{read_genotypes}
\alias{read_genotypes}
\title{Reads raw genotype files from PLINK and combines them into a single genotype matrix.}
\usage{
read_genotypes(files)
}
\arguments{
\item{files}{A vector of filenames to read}
}
\value{
A \link[data.table:data.table]{data.table::data.table} of dosages for individuals (rows) and SNPs (columns).
Individuals that are missing from some files will have their dosages set to NA for the SNPs in these files.
}
\description{
Reads raw genotype files from PLINK and combines them into a single genotype matrix.
}
