% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobitGwas.R
\name{regress_snp}
\alias{regress_snp}
\title{Run Tobit Regression on a single SNP}
\usage{
regress_snp(geno, pheno, covars, detection_limit, hde_test = FALSE)
}
\arguments{
\item{geno}{A numeric vector of dosages of one SNP for all individuals}

\item{pheno}{A numeric vector of phenotypes for all individuals.
Measurements below detection limit must be set to a value less than or equal detection limit.}

\item{covars}{N x M covariate matrix (rows are individuals)}

\item{detection_limit}{The detection limit of the phenotype}

\item{hde_test}{If TRUE, a test for the Hauck-Donner effect will be performed.
The result is currently not returned}
}
\value{
A named vector containing estimate, standard error, z and p values of the regression.
}
\description{
Runs a Tobit-I censored regression on the provided SNP.
}
